cambio de variable de float a decimal y cambio en la hoja de estilo para buscar codigo de producto

+++ database/migrations/2024_02_19_222336_create_productos_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->string('codigo',40);
             $table->string('nombre', 40);
             $table->string('detalle',50)->nullable();
-            $table->float('precio');
+            $table->decimal('precio', 8, 2);
             $table->integer('stock');
             $table->longText('imagen')->nullable();
             $table->timestamps();
